Refactoring and code coverage
