+++ Scd.ProjectX.Client/Scd.ProjectX.Client.Example/Observers/UserTradeObserver.cs
@@ -6,6 +6,7 @@ namespace Scd.ProjectX.Client.Example.Observers
     public class UserTradeObserver : IObserver<UserTradeEvent>
     {
         private int _throughputCounter = 0;
+
         public void OnCompleted()
         {
             Console.WriteLine($"onCompleted: {nameof(UserTradeObserver)}");
