+++ Scd.ProjectX.Client/Scd.ProjectX.Client.Example/appsettings.json
