+++ Scd.ProjectX.Client/Scd.ProjectX.Client/Messaging/Dispatchers/UserAccountDispatcher.cs
@@ -7,15 +7,13 @@ namespace Scd.ProjectX.Client.Messaging.Dispatchers
     /// Initializes a new instance of the <see cref="UserAccountDispatcher"/> class.
     /// </summary>
     /// <param name="connection">The <see cref="HubConnection"/>.</param>
-    public class UserAccountDispatcher :   
+    public class UserAccountDispatcher :
         EventDispatcher<UserAccountEvent>,
         IEventDispatcher<UserAccountEvent>
     {
         public UserAccountDispatcher(HubConnection connection)
             : base(connection, "GatewayUserAccount", "UnsubscribeAccounts")
         {
-
         }
-
-    }   
+    }
 }
