+++ Scd.ProjectX.Client/Scd.ProjectX.Client/Rest/IMarketDataFacade.cs
@@ -13,7 +13,7 @@ namespace Scd.ProjectX.Client.Rest
         /// </summary>
         /// <param name="request">The request object.</param>
         Task<List<Candle>> GetBars(BarsRequest request);
-        
+
         /// <summary>
         /// Retrieves historical bars (candles) for a specific contract within a given time range.
         /// </summary>
@@ -27,7 +27,7 @@ namespace Scd.ProjectX.Client.Rest
         /// <param name="includePartialBar">Whether to include a partial bar representing the current time unit.</param>
         /// <returns>The matching bars.</returns>
         Task<List<Candle>> GetBars(string contractId, DateTime startTime, DateTime endTime, int unitLength = 5, Unit unit = Unit.Minute, int limit = 100, bool live = true, bool includePartialBar = false);
-        
+
         /// <summary>
         /// Retrieves a list of contracts based on the provided search request.
         /// </summary>
@@ -43,6 +43,6 @@ namespace Scd.ProjectX.Client.Rest
         /// <param name="query">Search text</param>
         /// <param name="live">Whether to retrieve bars using the sim or live data subscription.</param>
         /// <returns>The matching contracts.</returns>
-        Task<List<Contract>> GetContracts(string contractId ="", string query = "", bool live = true);
+        Task<List<Contract>> GetContracts(string contractId = "", string query = "", bool live = true);
     }
 }
