+++ Scd.ProjectX.Client/Scd.ProjectX.Client/Utility/ProjectXClientException.cs
